+++ src/typeguard/_transformer.py
@@ -410,8 +410,11 @@ class AnnotationTransformer(NodeTransformer):
                 self.generic_visit(node)
 
                 # If the transformer erased the slice entirely, just return the node
-                # value without the subscript
-                if sys.version_info >= (3, 9) and not hasattr(node, "slice"):
+                # value without the subscript (unless it's Optional, in which case erase
+                # the node entirely
+                if self._memo.name_matches(node.value, "typing.Optional"):
+                    return None
+                elif sys.version_info >= (3, 9) and not hasattr(node, "slice"):
                     return node.value
                 elif sys.version_info < (3, 9) and not hasattr(node.slice, "value"):
                     return node.value